Move Node detection Stream functionality from primary entry point (core) to Node (only) entry point
