+++ index.d.ts
@@ -4,14 +4,23 @@ Typings for Node.js specific entry point.
 
 import type {Readable as NodeReadableStream} from 'node:stream';
 import type {ReadableStream as WebReadableStream} from 'node:stream/web';
-import type {FileTypeResult} from './core.js';
+import type {FileTypeResult, StreamOptions} from './core.js';
 import {FileTypeParser} from './core.js';
+
+export type ReadableStreamWithFileType = NodeReadableStream & {
+	readonly fileType?: FileTypeResult;
+};
 
 export declare class NodeFileTypeParser extends FileTypeParser {
 	/**
 	@param stream - Node.js `stream.Readable` or Web API `ReadableStream`.
 	*/
 	fromStream(stream: WebReadableStream<Uint8Array> | NodeReadableStream): Promise<FileTypeResult | undefined>;
+
+	/**
+	Works the same way as {@link fileTypeStream}, additionally taking into account custom detectors (if any were provided to the constructor).
+	 */
+	toDetectionStream(readableStream: NodeReadableStream, options?: StreamOptions): Promise<ReadableStreamWithFileType>;
 }
 
 /**
@@ -25,5 +34,36 @@ The file type is detected by checking the [magic number](https://en.wikipedia.or
 export function fileTypeFromFile(path: string): Promise<FileTypeResult | undefined>;
 
 export function fileTypeFromStream(stream: WebReadableStream<Uint8Array> | NodeReadableStream): Promise<FileTypeResult | undefined>;
+
+/**
+Returns a `Promise` which resolves to the original readable stream argument, but with an added `fileType` property, which is an object like the one returned from `fileTypeFromFile()`.
+
+This method can be handy to put in between a stream, but it comes with a price.
+Internally `stream()` builds up a buffer of `sampleSize` bytes, used as a sample, to determine the file type.
+The sample size impacts the file detection resolution.
+A smaller sample size will result in lower probability of the best file type detection.
+
+**Note:** This method is only available when using Node.js.
+**Note:** Requires Node.js 14 or later.
+
+@param readableStream - A [readable stream](https://nodejs.org/api/stream.html#stream_class_stream_readable) containing a file to examine.
+@returns A `Promise` which resolves to the original readable stream argument, but with an added `fileType` property, which is an object like the one returned from `fileTypeFromFile()`.
+
+@example
+```
+import got from 'got';
+import {fileTypeStream} from 'file-type';
+
+const url = 'https://upload.wikimedia.org/wikipedia/en/a/a9/Example.jpg';
+
+const stream1 = got.stream(url);
+const stream2 = await fileTypeStream(stream1, {sampleSize: 1024});
+
+if (stream2.fileType?.mime === 'image/jpeg') {
+	// stream2 can be used to stream the JPEG image (from the very beginning of the stream)
+}
+```
+ */
+export function fileTypeStream(readableStream: NodeReadableStream, options?: StreamOptions): Promise<ReadableStreamWithFileType>;
 
 export * from './core.js';
