+++ index.js
@@ -4,7 +4,7 @@ Node.js specific entry point.
 
 import {ReadableStream as WebReadableStream} from 'node:stream/web';
 import * as strtok3 from 'strtok3';
-import {FileTypeParser} from './core.js';
+import {FileTypeParser, reasonableDetectionSizeInBytes} from './core.js';
 
 export class NodeFileTypeParser extends FileTypeParser {
 	async fromStream(stream) {
@@ -14,6 +14,41 @@ export class NodeFileTypeParser extends FileTypeParser {
 		} finally {
 			await tokenizer.close();
 		}
+	}
+
+	async toDetectionStream(readableStream, options = {}) {
+		const {default: stream} = await import('node:stream');
+		const {sampleSize = reasonableDetectionSizeInBytes} = options;
+
+		return new Promise((resolve, reject) => {
+			readableStream.on('error', reject);
+
+			readableStream.once('readable', () => {
+				(async () => {
+					try {
+						// Set up output stream
+						const pass = new stream.PassThrough();
+						const outputStream = stream.pipeline ? stream.pipeline(readableStream, pass, () => {}) : readableStream.pipe(pass);
+
+						// Read the input stream and detect the filetype
+						const chunk = readableStream.read(sampleSize) ?? readableStream.read() ?? new Uint8Array(0);
+						try {
+							pass.fileType = await this.fromBuffer(chunk);
+						} catch (error) {
+							if (error instanceof strtok3.EndOfStreamError) {
+								pass.fileType = undefined;
+							} else {
+								reject(error);
+							}
+						}
+
+						resolve(outputStream);
+					} catch (error) {
+						reject(error);
+					}
+				})();
+			});
+		});
 	}
 }
 
@@ -31,4 +66,8 @@ export async function fileTypeFromStream(stream, fileTypeOptions) {
 	return (new NodeFileTypeParser(fileTypeOptions)).fromStream(stream);
 }
 
-export {fileTypeFromBuffer, fileTypeFromBlob, fileTypeStream, FileTypeParser, supportedMimeTypes, supportedExtensions} from './core.js';
+export async function fileTypeStream(readableStream, options = {}) {
+	return new NodeFileTypeParser().toDetectionStream(readableStream, options);
+}
+
+export {fileTypeFromBuffer, fileTypeFromBlob, FileTypeParser, supportedMimeTypes, supportedExtensions} from './core.js';
